% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polychart.R
\name{polychart}
\alias{polychart}
\title{Create polychart-based scatter plot}
\usage{
polychart(gg_obj, width = NULL, height = NULL, data_cols = c(0, 1),
  palette = c("#3182bd", "#fd8d3c", "#74c476"))
}
\description{
More info here.
}

