% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polychart.R
\name{renderPolychart}
\alias{renderPolychart}
\title{Widget render function for use in Shiny}
\usage{
renderPolychart(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

